Compilation of a sample Xenomai program from within QEMU
Thanks to the help of a maintainer of the Intel Xenomai repository, it was finally
possible to run successfully the latency test.
After that, it was simple enough to compile and execute a demo program written
by the original Xenomai author.
The next step is to see if this process may be automated with Yocto.